Add custom configuration for Termit terminal emulator
This configuration matches our established preferences and mostly
mirrors the configuration we were using with Terminator.

One exception is that the cursor is shown in the text foreground color
rather than in `#deb887`, which is tolerable.  It doesn't appear that
Termit(/VTE?) allows for setting this.

In this commit we also added support for chromium-style tab flipping.
We were already using page-up/page-down for moving tabs around; we
might as well support using them for changing tabs as well.

We had wanted to move away from Terminator for awhile; it had some
nasty habits such as writing data unnecessarily to /tmp and it was
just bigger than we preferred.  However we were driven away by the
update to v1.90 which added a light-colored border (presumably a bug)
around the title that was impossible to remove, and showed the
scrollbars when adding new tabs even when scrollbars were explicitly
disabled in the profile.

Having a lua-based scripting/configuration language for our terminal
is perfect, and with Termit we managed to achieve the desired UI
minimalism.